Moved some Quote files to GlyssenEngine
