+++ Glyssen/BookSetUtils.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using L10NSharp;
 using SIL;
 using SIL.DblBundle.Text;
 using SIL.Scripture;
