+++ Glyssen/Dialogs/QuotationMarksDlg.cs
@@ -8,11 +8,11 @@ using Glyssen.Character;
 using Glyssen.Controls;
 using Glyssen.Paratext;
 using Glyssen.Properties;
-using Glyssen.Quote;
 using Glyssen.Shared;
 using Glyssen.Utilities;
 using GlyssenEngine.Bundle;
 using GlyssenEngine.Paratext;
+using GlyssenEngine.Quote;
 using L10NSharp.TMXUtils;
 using L10NSharp.UI;
 using SIL;
