+++ Glyssen/Project.cs
@@ -19,10 +19,10 @@ using Glyssen.Properties;
 using Glyssen.Quote;
 using Glyssen.Shared;
 using Glyssen.Shared.Bundle;
-using Glyssen.Utilities;
 using GlyssenEngine;
 using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
+using GlyssenEngine.Quote;
 using GlyssenEngine.Utilities;
 using GlyssenEngine.VoiceActor;
 using Paratext.Data;
