+++ Glyssen/Properties/Resources.resx
@@ -127,9 +127,6 @@
   <data name="CharacterVerseData" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\CharacterVerse.txt;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
-  <data name="QuoteSystemData" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\quoteSystems.xml;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;utf-8</value>
-  </data>
   <data name="112_Plus_Green_16x16_72" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\112_Plus_Green_16x16_72.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
