+++ Glyssen/Quote/ParserCharacterRepository.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using Glyssen.Character;
 using Glyssen.Shared;
 using GlyssenEngine.Character;
 using SIL.Extensions;
