+++ GlyssenEngine/GlyssenEngine.csproj
@@ -25,6 +25,7 @@
     <None Remove="Resources\Annotations.txt" />
     <None Remove="Resources\BiblicalAuthors.txt" />
     <None Remove="Resources\CharacterDetail.txt" />
+    <None Remove="Resources\quoteSystems.xml" />
     <None Remove="Resources\RelatedCharacters.xml" />
   </ItemGroup>
 
@@ -81,6 +82,7 @@
     <Content Include="Resources\Annotations.txt" />
     <Content Include="Resources\BiblicalAuthors.txt" />
     <Content Include="Resources\CharacterDetail.txt" />
+    <Content Include="Resources\quoteSystems.xml" />
     <Content Include="Resources\RelatedCharacters.xml" />
   </ItemGroup>
