+++ GlyssenEngine/Resources.resx
@@ -127,6 +127,9 @@
   <data name="CharacterDetail" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Resources\CharacterDetail.txt;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
+  <data name="QuoteSystemData" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Resources\quoteSystems.xml;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;utf-8</value>
+  </data>
   <data name="RelatedCharacters" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Resources\RelatedCharacters.xml;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
