+++ GlyssenEngine/Resources/quoteSystems.xml
