+++ GlyssenTests/ProjectTests.cs
@@ -7,13 +7,13 @@ using Glyssen;
 using Glyssen.Bundle;
 using Glyssen.Character;
 using Glyssen.Paratext;
-using Glyssen.Quote;
 using Glyssen.Shared;
 using Glyssen.Shared.Bundle;
 using GlyssenEngine;
 using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
 using GlyssenEngine.Paratext;
+using GlyssenEngine.Quote;
 using GlyssenTests.Bundle;
 using NUnit.Framework;
 using SIL.DblBundle.Text;
