+++ GlyssenTests/TestProject.cs
@@ -7,12 +7,12 @@ using System.Threading;
 using System.Xml;
 using Glyssen;
 using Glyssen.Bundle;
-using Glyssen.Quote;
 using Glyssen.Shared;
 using Glyssen.Shared.Bundle;
 using GlyssenEngine;
 using GlyssenEngine.Bundle;
 using GlyssenEngine.Character;
+using GlyssenEngine.Quote;
 using SIL.DblBundle.Text;
 using SIL.DblBundle.Usx;
 using SIL.IO;
