+++ RefTextDevUtilities/ReferenceTextUtility.cs
@@ -7,9 +7,9 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using Glyssen.Character;
-using Glyssen.Quote;
 using Glyssen.Shared;
 using GlyssenEngine.Character;
+using GlyssenEngine.Quote;
 using OfficeOpenXml;
 using SIL.Reflection;
 using SIL.Scripture;
